packages/gaebdb/pyproject.toml
[project]
name = "gaebdb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Leon Weimann", email = "nachhilfe@leonweimann.de" }]
requires-python = ">=3.13"
dependencies = [
    "psycopg[binary]>=3.2.10",
    "python-dotenv>=1.1.1",
    "sqlalchemy[asyncio]>=2.0.44",
]

[build-system]
requires = ["uv_build>=0.8.21,<0.9.0"]
build-backend = "uv_build"
